Generate a dataframe of diagnostic performance over the range of thresholds specified in the main dx function call.

Generate a dataframe of diagnostic performance over the range of thresholds specified in the main dx function call.

dx_diagnostic_table(
  dx_obj,
  includeFractions = TRUE,
  includeAUC = TRUE,
  includeAccuracy = TRUE,
  includeSensitivity = TRUE,
  includeSpecificity = TRUE,
  includePPV = TRUE,
  includeNPV = TRUE,
  includeOR = TRUE,
  includeF1 = TRUE
)

Arguments

dx_obj

An object of class dx

includeFractions

Logical for inclusion of the fractions appended to the estimates and CIs

includeAUC

Logical to include AUC in data frame

includeAccuracy

Logical to include Accuracy in data frame

includeSensitivity

Logical to include sensitivity in data frame

includeSpecificity

Logical to include Specificity in data frame

includePPV

Logical to include Positive predictive value in data frame

includeNPV

Logical to include Negative predictive value in data frame

includeOR

Logical to include Odds Ratio in data frame

includeF1

Logical to include F1 in data frame

Examples

dx_obj <- dx(
  data = dx_heart_failure,
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  threshold_range = c(.1, .2, .3),
  setthreshold = .3,
  grouping_variables = c("AgeGroup", "Sex", "AgeSex")
)
#> Error in tn_1[, 1]: incorrect number of dimensions
dx_diagnostic_table(dx_obj, includeOR = FALSE, includeF1 = FALSE)
#> Error in dplyr::filter(., variable == "Overall"): object 'dx_obj' not found