Heart attack outcomes and predictions

A data set containing heart attack outcomes for 261 patients along with predicted probabilities of having a heart attack.

dx_heart_failure

Format

A data frame with 261 rows and 5 variables:

AgeGroup

Age group

Sex

sex

truth

Heart failure (outcome)

predicted

Predicted outcome from model

AgeSex

Age and sex group

Source

https://www.kaggle.com/imnikhilanand/heart-attack-prediction