Generate a dataframe of diagnostic performance over the range of thresholds specified in the main dx function call.

Generate a dataframe of diagnostic performance over the range of thresholds specified in the main dx function call.

dx_diagnostic_table(
  dx_obj,
  includeFractions = TRUE,
  includeAUC = TRUE,
  includeAccuracy = TRUE,
  includeSensitivity = TRUE,
  includeSpecificity = TRUE,
  includePPV = TRUE,
  includeNPV = TRUE,
  includeOR = TRUE,
  includeF1 = TRUE
)

Arguments

dx_obj
An object of class dx

includeFractions
Logical for inclusion of the fractions appended to the estimates and CIs

includeAUC
Logical to include AUC in data frame

includeAccuracy
Logical to include Accuracy in data frame

includeSensitivity
Logical to include sensitivity in data frame

includeSpecificity
Logical to include Specificity in data frame

includePPV
Logical to include Positive predictive value in data frame

includeNPV
Logical to include Negative predictive value in data frame

includeOR
Logical to include Odds Ratio in data frame

includeF1
Logical to include F1 in data frame

Examples

dx_obj <- dx(
  data = dx_heart_failure,
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  threshold_range = c(.1, .2, .3),
  setthreshold = .3,
  grouping_variables = c("AgeGroup", "Sex", "AgeSex")
)
dx_diagnostic_table(dx_obj, includeOR = FALSE, includeF1 = FALSE)

#>   threshold                  AUC                     Accuracy
#> 1       0.1 0.904 (0.864, 0.943) 67.8% (61.8%, 73.4%) 177/261
#> 2       0.2 0.904 (0.864, 0.943) 77.0% (71.4%, 82.0%) 201/261
#> 3       0.3 0.904 (0.864, 0.943) 79.3% (73.9%, 84.1%) 207/261
#>                  Sensitivity                  Specificity
#> 1 95.9% (89.9%, 98.9%) 94/98  50.9% (43.0%, 58.8%) 83/163
#> 2 89.8% (82.0%, 95.0%) 88/98 69.3% (61.6%, 76.3%) 113/163
#> 3 84.7% (76.0%, 91.2%) 83/98 76.1% (68.8%, 82.4%) 124/163
#>     Positive Predictive Value    Negative Predictive Value
#> 1 54.0% (46.3%, 61.6%) 94/174   95.4% (88.6%, 98.7%) 83/87
#> 2 63.8% (55.2%, 71.8%) 88/138 91.9% (85.6%, 96.0%) 113/123
#> 3 68.0% (59.0%, 76.2%) 83/122 89.2% (82.8%, 93.8%) 124/139